+++ docs/CHANGELOG.md
@@ -32,6 +32,7 @@ version 2.0.4.1
 * Add api method output.error and modify output.fail
 * Optimization results show
 * Fix bug in lib/core/poc.py: output.error is not tuple type
+* Update ZoomEye API address
 
 
 version 3.0